Fix up readme
This changes markdown style to be more consistent with the usual
markdown style. Also, importantly, this changes the plugin license
requirements to require an OSI-approved license instead of WTFPL.
WTFPL is basically public domain, and doesn't work in a number of
countries. You can't really use it for serious projects.